+++ cmd/attach.go
@@ -1,10 +1,14 @@
 package cmd
 
 import (
+	"fmt"
+	"strings"
+
 	"github.com/spf13/cobra"
 )
 
 func newAttachCommand() *cobra.Command {
+	index := 1
 	var cmd = &cobra.Command{
 		Use:   "attach",
 		Short: "Attach local stdio to a running container",
@@ -13,9 +17,17 @@ func newAttachCommand() *cobra.Command {
 		// TODO: ArgsInUseLine: "service"
 		PreRun: configSavePreRun,
 		RunE: func(cmd *cobra.Command, args []string) error {
-			cid, err := dockerContainerID(args[0])
+			service := args[0]
+			cid, err := dockerContainerID(service)
 			if err != nil {
 				return err
+			}
+
+			lines := strings.Split(cid, "\n")
+			if len(lines) >= index {
+				cid = lines[index-1]
+			} else {
+				return fmt.Errorf("Index %d not found for service %s", index, service)
 			}
 
 			cmdArgs := []string{"attach"}
@@ -28,6 +40,11 @@ func newAttachCommand() *cobra.Command {
 			)
 		},
 	}
+
+	cmd.Flags().SortFlags = false
+
+	cmd.Flags().IntVarP(&index, "index", "", 1, "index of the container if there are multiple\ninstances of a service")
+	cmd.Flags().SetAnnotation("index", "muss-only", []string{"true"})
 
 	cmd.Flags().StringP("detach-keys", "", "ctrl-c", "Override the key sequence for detaching a container")
 	cmd.Flags().BoolP("no-stdin", "", false, "Do not attach STDIN")