+++ cmd/attach_test.go
@@ -1,6 +1,7 @@
 package cmd
 
 import (
+	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -40,6 +41,44 @@ attach
 --no-stdin=false
 --sig-proxy=false
 some:foo:cid
+`
+
+			assert.Nil(t, err)
+			assert.Equal(t, "", stderr)
+			assert.Equal(t, expOut, stdout)
+		})
+
+		t.Run("index", func(*testing.T) {
+			os.Setenv("MUSS_TEST_PS_SCALE", "1")
+			defer os.Unsetenv("MUSS_TEST_PS_SCALE")
+
+			stdout, stderr, err := testCmdBuilder(newAttachCommand, []string{"foo", "--index=1"})
+
+			expOut := `docker
+attach
+--detach-keys=ctrl-c
+--no-stdin=false
+--sig-proxy=false
+some:foo:cid
+`
+
+			assert.Nil(t, err)
+			assert.Equal(t, "", stderr)
+			assert.Equal(t, expOut, stdout)
+		})
+
+		t.Run("index=2", func(*testing.T) {
+			os.Setenv("MUSS_TEST_PS_SCALE", "1")
+			defer os.Unsetenv("MUSS_TEST_PS_SCALE")
+
+			stdout, stderr, err := testCmdBuilder(newAttachCommand, []string{"foo", "--index", "2"})
+
+			expOut := `docker
+attach
+--detach-keys=ctrl-c
+--no-stdin=false
+--sig-proxy=false
+second:foo:cid
 `
 
 			assert.Nil(t, err)